+++ source/Examples/Resources/SqlDatabaseRole/6-MembersToIncludeAndExcludeInDatabaseRole.ps1
@@ -27,8 +27,8 @@ Configuration Example
         {
             ServerName           = 'sqltest.company.local'
             InstanceName         = 'DSC'
-            Database             = 'AdventureWorks'
-            Name                 = 'ReportViewer'
+            DatabaseName         = 'AdventureWorks'
+            RoleName             = 'ReportViewer'
             MembersToInclude     = @('CONTOSO\Barbara', 'CONTOSO\Fred')
             MembersToExclude     = @('CONTOSO\Intern1')
             Ensure               = 'Present'